+++ src/com/alienation/coregamefiles/gamefunctionclasses/MoveRoom.java
@@ -1,0 +1,28 @@
+package com.alienation.coregamefiles.gamefunctionclasses;
+
+import com.alienation.coregamefiles.enums.Rooms;
+
+import static com.alienation.coregamefiles.charactersetc.Player.setPreviousRoom;
+import static com.alienation.coregamefiles.charactersetc.Player.setTempRoom;
+import static com.alienation.coregamefiles.gameart.TextColors.ANSI_RED;
+import static com.alienation.coregamefiles.gameart.TextColors.ANSI_RESET;
+import static com.alienation.coregamefiles.gamefunctionclasses.LoadRoom.loadRoom;
+import static com.alienation.coregamefiles.gamefunctionclasses.Menu.displayMenu;
+import static com.alienation.coregamefiles.hashmaps.DirectionsHashMap.getAvailableDirectionsMap;
+
+public class MoveRoom {
+
+    // Move Room from one to another
+    public static void moveRoom(String direction, Rooms currentRoom) throws Exception {
+        Rooms nextRoom = getAvailableDirectionsMap().get(currentRoom).get(direction);
+        if(nextRoom != null){
+            setPreviousRoom(currentRoom);
+            setTempRoom(currentRoom);
+            loadRoom(nextRoom);
+        }
+        else{
+            System.out.println(ANSI_RED + "\nThere doesn't seem to be a door this way.\n" + ANSI_RESET);
+            displayMenu();
+        }
+    }
+}
