+++ src/com/alienation/coregamefiles/gamefunctionclasses/ReadThings.java
@@ -1,0 +1,11 @@
+package com.alienation.coregamefiles.gamefunctionclasses;
+
+public class ReadThings {
+
+    //read clues
+    public static void read(){
+        System.out.println("Can't Read yet!!");
+        //TODO: Read
+        //
+    }
+}
