+++ src/com/alienation/coregamefiles/rooms/AlienRoom.java
@@ -30,6 +30,7 @@ import java.util.List;
 
 import static com.alienation.coregamefiles.gameart.TextColors.ANSI_BLUE;
 import static com.alienation.coregamefiles.gameart.TextColors.ANSI_RESET;
+import static com.alienation.coregamefiles.gamefunctionclasses.AlienAttack.getAttackCount;
 import static com.alienation.coregamefiles.hashmaps.AvailableItemsHashMap.getAvailableItemsMap;
 
 
@@ -53,7 +54,7 @@ public class AlienRoom extends Room {
         List<String> availableItems = getAvailableItemsMap().get(Rooms.AlienRoom);
         if (!availableItems.contains("Humanoid")) {
             return getLastStory();
-        }else if (Menu.attackCount > 0 && availableItems.contains("Humanoid")){
+        }else if (getAttackCount() > 0 && availableItems.contains("Humanoid")){
                 return getUpdatedStory();
         }else{
             return getInitialStory();
